Refactor shared timeline visibility logic
